# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "同桌小程序",
        "title": "Deskmat",
        "termsOfService": "http://swagger.io/terrms/",
        "contact": {
            "name": "qianren",
            "email": "1911401642@qq.com"
        },
        "version": "1.0.0"
    },
    "host": "119.3.2.168:4016",
    "basePath": "/api/v1/",
    "paths": {
...
        "/card": {
            "get": {
                "description": "\"获取自己的名片信息\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card"
                ],
                "summary": "我的名片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Card"
                        }
                    },
                    "401": {
                        "description": "身份验证失败"
                    },
                    "404": {
                        "description": "获取失败"
                    }
                }
            },
            "put": {
                "description": "\"修改名片的昵称，标签，头像，用户宣言\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card"
                ],
                "summary": "修改名片信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "名片信息",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Cardinfor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功"
                    },
                    "400": {
                        "description": "修改失败"
                    },
                    "401": {
                        "description": "验证失败"
                    }
                }
            },
            "post": {
                "description": "\"设置名片的昵称，标签，头像，用户宣言\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card"
                ],
                "summary": "设置名片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "名片信息",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Cardinfor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "设置成功"
                    },
                    "400": {
                        "description": "设置失败"
                    },
                    "401": {
                        "description": "身份验证失败 重新登录"
                    }
                }
            }
        },
        "/card/avatar": {
            "post": {
                "description": "\"修改名片头像\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card"
                ],
                "summary": "修改头像",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"mgs\":\"success\"}",
                        "schema": {
                            "$ref": "#/definitions/model.Card"
                        }
                    },
                    "400": {
                        "description": "上传失败,请检查token与其他配置参数是否正确"
                    }
                }
            }
        },
...
        "/dailyrecord/end": {
            "put": {
                "description": "\"强制中断当前同桌\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dailyrecord"
                ],
                "summary": "强制中断同桌",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "强制中断成功"
                    },
                    "400": {
                        "description": "强制中断关系失败"
                    },
                    "401": {
                        "description": "身份验证失败 重新登录"
                    },
                    "500": {
                        "description": "用户更新状态失败"
                    }
                }
            }
        },
...
        "/dailyrecord/send": {
            "post": {
                "description": "\"发送一条消息\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dailyrecord"
                ],
                "summary": "发送消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "每日打卡的内容",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Record"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "发送成功"
                    },
                    "400": {
                        "description": "发送失败"
                    },
                    "401": {
                        "description": "身份验证失败 重新登录"
                    },
                    "500": {
                        "description": "服务器发生错误"
                    }
                }
            }
        },
        "/dailyrecord/update": {
            "put": {
                "description": "\"每日更新打卡天数\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dailyrecord"
                ],
                "summary": "更新打卡",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功"
                    },
                    "400": {
                        "description": "更新失败，同桌关系已经解除"
                    },
                    "401": {
                        "description": "身份验证失败 重新登录"
                    },
                    "500": {
                        "description": "失败"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Card": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "college": {
                    "description": "学院",
                    "type": "string"
                },
                "declaration": {
                    "type": "string"
                },
                "grade": {
                    "description": "年级",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "infor": {
                    "type": "string"
                },
                "major": {
                    "description": "专业",
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "sha": {
                    "type": "string"
                },
                "status": {
                    "description": "记录是否有同桌",
                    "type": "string"
                },
                "tag1": {
                    "type": "string"
                },
                "tag2": {
                    "type": "string"
                },
                "tag3": {
                    "type": "string"
                },
                "tag4": {
                    "type": "string"
                },
                "tag5": {
                    "type": "string"
                },
                "users_id": {
                    "type": "string"
                }
            }
        },
        "model.Cardinfor": {
            "type": "object",
            "properties": {
                "college": {
                    "type": "string"
                },
                "declaration": {
                    "type": "string"
                },
                "grade": {
                    "type": "string"
                },
                "infor": {
                    "type": "string"
                },
                "major": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "tag1": {
                    "type": "string"
                },
                "tag2": {
                    "type": "string"
                },
                "tag3": {
                    "type": "string"
                },
                "tag4": {
                    "type": "string"
                },
                "tag5": {
                    "type": "string"
                }
            }
        },
...
        "model.Record": {
            "type": "object",
            "properties": {
                "information": {
                    "type": "string"
                }
            }
        },
...
    }
}